requirements.txt
flask==2.3.3
opencv-python==4.8.1.78
face-recognition==1.3.0
psycopg2-binary==2.9.7
pillow==10.0.1
requests==2.31.0
python-dateutil==2.8.2